Changed image path so images now stop giving us the 404 error. Changed tostring method so it gives a better view of the board.
Added getCell method to Gameboard that allows me to get a cell with
whatever coordinate i send it.

fixed generatemap so it actually generates a gameboard full of cells
instead of a gameboard that is empty.  

added GraphicsPanel that will hold all of our cool drawings and stuff.

fixed GUI main panel to make it look a bit better (still needs more
work) 

added down and left button action that adds terrain onto the gameboard
as well as updates the drawing.